requirements.txt
requests
python-whois
phonenumbers
exifread
reportlab
tabulate